Try to move "dangling" rows in upgradedb
Instead of failing loudly for invalid records (which happens way too
often), this attempts to move those offending data to another table and
carry on with the migration if possible. This table for dangling data
are copied with CREATE TABLE ... AS SELECT ... and could miss some
indexing and stuff, but this is only meant for temporary storage, so
this is probably not a big deal. If copying went well, the dangling
data are automatically deleted so we can carry on with migration.

Additionally, this commit removes the upgrade check on TaskFail, and
added check on TaskReschedule. This is because TaskFail is not actually
being migrated in 2.2, while TaskReschedule is, and we concluded this
is likely a typo during implementation and not an intentional choice.

(cherry picked from commit 277abdc403424c13a9adb4bb2a5f74f25ccff8e3)