Abstract task group children serialization
This introduces a new function 'serialize_for_task_group' to DAGNode,
which should implement serialization logic when a DAGNode is put into a
task group. Since all nodes belonging to a DAG are either the root
group, or always in a parent task group, this guarantees a DAG's all
nodes can be correctly serialized.