Deprecate the `extra__...` prefix in connection extras
Previously, connection "extra" fields which were added as custom fields in the
webserver connection form had to be named with prefix `extra__<conn_type>__`.
This was because custom fields are registered globally on the connection view model,
so the prefix was necessary to prevent collisions.

But the prefix is ugly and cumbersome in the `extra` field.  So now what we do is
add this prefix when defining the field internally in the model, and strip it when
saving the connection.