Implement dagrun notiification system for airflow
This commit introduces a new feature in airflow that allows users to create notifications
for when a dagrun enters a particular state. This can be useful for keeping track of the
status of important dags and receiving alerts when they fail or there are changes in their execution.

To use this feature, users can simply specify the desired dagrun states and a notification method
(e.g. email,) when creating a new notification. Optionally, users can also specify certain dag tags to
only monitor the specific dags' dagrun state. The notification will then be triggered whenever
 the dagrun enters the specified state.

Note that the current implementation only includes email as a notification method, but this can be
expanded in the future to include other methods such as slack or pagerduty.