Remove execution_date and logical_date from arguments where function/API is used to look up a DAG run
Remove execution_date and logical_date from arguments where function/API is used to look up a DAG run

- Resolve compatibility issues and refactor `execution_date` to `logical_date`
- Resolve compatibility tests
- Correct import paths after rebase
- Address static checks
- Refactor `execution_date` to `logical_date`
- Add missing DAG files for tests
- Enhance GCP ML and CloudBuild tests using helpers for compatibility
- Fix mypy errors
- Miscellaneous fixes and removals of `execution_date`
- Resolve compatibility tests
- Correct import paths after rebase
- Address static checks
- Refactor `execution_date` to `logical_date`
- Add missing DAG files for tests
- Enhance GCP ML and CloudBuild tests using helpers for compatibility
- Fix mypy errors
- Miscellaneous fixes and removals of `execution_date`
- Resolve compatibility tests
- Correct import paths after rebase
- Address static checks
- Refactor `execution_date` to `logical_date`
- Add missing DAG files for tests
- Enhance GCP ML and CloudBuild tests using helpers for compatibility
- Fix mypy errors
- Miscellaneous fixes and removals of `execution_date`
- Drop execution_date unique constraint on DagRun
- The column has also been renamed to logical_date, although the Python
model is not changed. This allows us to not need to fix all the Python
code at once (we'll do that later), but still do the two changes in one
migration instead of two.
- Use test helpers in GCP MLEngine tests for compat
- Using Airflow internals directly presents a problem when dealing with
compatibility in tests (since the same tests must run against Airflow 2
and 3). The helpers already handle this well, so we should use them.
- Use test helpers in GCP CloudBuild tests for compat
- Using Airflow internals directly presents a problem when dealing with
compatibility in tests (since the same tests must run against Airflow 2
and 3). The helpers already handle this well, so we should use them.
- Mark db tests
- Some compat code to make DAG.clear() still work
- Remove unneeded test cases for compat code
- Use test helpers in GCS-BQ tests for compat
- Using Airflow internals directly presents a problem when dealing with
compatibility in tests (since the same tests must run against Airflow 2
and 3). The helpers already handle this well, so we should use them.