Add logging to the migration commands
This PR enhances logging for the migration commands and also
ensures that alembic doesn't change the logger when logging.

A PR was raised to address the issue of migration changing logging,
see f1664674d859a262e93fb3110557a1e71138ca8b, but that did address it
for other commands. This PR addresses it for all the migration commands