Allow logical_date=None DAG-triggering operators
Previously, TriggerDagRunOperator heavily relies on the unique non-null
property of logical_date. It uses logical date to track what DAG run is
being created (via a trigger), and use the logical date to check if the
run is triggered correctly.

This becomes a problem when logical date can be None because that None
value cannot be used to track the run. Fortunately, DAG runs already
have a unique tracking ID (run_id). We can just use it instead.

The change however required the entire DAG- and task-triggering
mechanism to be changed since everything in the stack previously uses
logical date.