Change version handling in provider documentation to packaging version (#48249)
We have been using SemVer to handle version updates in provider
documentation breeze commands, which started to break at the moment
we started to have pre-release packages - Puthon packaging follows
PEP-440 and it is implemented in packaging.version module.

This module does not have all the nice features semver has (like
bumping the major/minor/patchlevel versions so we had to implement
it ourselves (packaging Versions are immutable and we need to
recreate the version after bump.