[astro] Reduce API server memory growth due to Task callback source in serialized dag (#58)
The _source_ of all callbacks was being included in the serialized dag JSON in
v1 and v2 of the format, and due to some other (not yet fully identified
issues) there ends up multiple copies of this json string in memory.

However as we noticed in building v3 we don't need this value at all, we just
need to know that it exist, so we replace it "in-memory" to be `[True]`, that
way it is much shorter but maintains the same structure, just in case
something depends on it being an array.