Add automatic support for pulling "XComArg"s in to task functions.
In order to support this it needs a change on the server side, to send
`task_args` as part of the TIRunContext argument. This needs to come from the
server because the source of this is in the dag file itself (which isn't
something Go has access to).

For this to actually it work it depends on an un-released change to the
execution API server (to add this field to the respons) and to the stub
decorator in the standard provider (so that we store the task args in a form
that is conducive to extracting this information. Currently it is stored as a
stringification of the python tuple!)

Also as part of this change we support passing other arguments (literals,
effectively) from the Python DAG to the Go function via the same mechanism.
There was an API decision to decide how to support this in Go, and for now, I
have chosen simple positional arguments. The other option would be a `Params`
struct. For example, something like this:

```python
@task.stub
def mytask(a, something_else): ...

with DAG():
  some_else = task1() # Get a result from some other task
  mytask(2, some_else) # And feed it to the go task, along with a literal
```

```go
// go
type Params struct {
  A int `mapstructure:"a"`
  SomethingElse *time.Time `mapstructure:"something_else"`
}
func mytask(ctx context.Context, params Params) error {
}
```

However _requiring_ that approach makes ever go task very verbose even if it
accepts only one or two parameters. So instead I have gone with the approach
of "all special params first" (like context, logger or sdk.Client -- the ones
detected by type) and then all task arguments next.

Meaning this would be invalid (logger must be before the `a` argument):

```go
func mytask(ctx context.Context, a int, logger *slog.Logger) error {
}
```